Rename `SimpleTrader` to `MadTrader`
